feat(Controllers): Add /accounts & /sessions routes
