+++ src/controllers/auth/index.ts
@@ -1,0 +1,2 @@
+export * from './AccountController'
+export * from './SessionController'
