+++ src/controllers/index.ts
@@ -1,6 +1,6 @@
-export * from './AuthController'
+export * from './auth'
+export * from './servers'
 export * from './UserController'
 export * from './MessageController'
 export * from './ChannelController'
 export * from './InviteController'
-export * from './servers'
